Extract Firestore read + write ops into host/firestore-{reads,writes}.ts
Move the getDoc/getDocs/count/aggregate/keyspace ops into
host/firestore-reads.ts and the setDoc/updateDoc/deleteDoc/addDoc/batchCommit/
txnCommit ops (with their write-payload prep, sentinel reconstruction, batch
applier, and txn read-set canonicalizer) into host/firestore-writes.ts, each
behind an is*Op predicate + handle*Op function.

handleOp now peels these families off with early-return routing after computing
the resolved db/lens preamble, so the eager db/lens resolution is preserved for
every op exactly as before. Case bodies and helpers moved verbatim; orphaned
pyric/firestore write imports dropped from host.ts. No behavior change.
